Fix #5782 Unclear error message on data conversion failure
