+++ molgenis-data-mapper/src/main/java/org/molgenis/data/mapper/service/impl/AlgorithmException.java
@@ -1,0 +1,9 @@
+package org.molgenis.data.mapper.service.impl;
+
+public class AlgorithmException extends RuntimeException
+{
+	public AlgorithmException(String message)
+	{
+		super(message);
+	}
+}
